WIP - manage helm dependencies
At this time, we can declare a service as a dependency. So it will not
be exported as deployment. We can use this kind of definition:
```
mariadb:
  image: mariadb
  labels:
    katenary.io/dependency: |
      name: mariadb-galera
      repository: http://...
      version: 7.x.x
      environment:
         foo: bar
```

It's not perfect at all, but it's a good starting point.

I need to rewrite the "depends_on" detection to use the right service
name that can be override. At this time, the default is
REALEASE_NAME-dep-name