Fix dependencies + add bats tests
Dependencies seems to work, there are still some problems to resolve and
many tests to add.

We decide to yse "bats" (submodule) to ensure that YAML files are well
generated with helm and to use "yq" command to query generated files.
It's less correct to check golang coverage, but it's a bit more easy to
ensure that Katenary works.